pyproject.toml
[project]
name = "phase3-django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "generic>=1.1.5",
    "generics>=7.0.0",
    "phonenumbers>=9.0.11",
    "pydantic>=2.11.7",
    "pymysql>=1.1.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=37.5.3",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]


